footagents-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langgraph==0.2.16
langchain==0.2.16
langchain-groq==0.1.9
pymongo==4.6.0
sentence-transformers==2.2.2
pydantic==2.5.0
python-dotenv==1.0.0
asyncio
python-multipart==0.0.6 